// src/app/interfaces/checkplanta.ts
export interface CheckPlanta{
    id?: number;	
    folio: number;	
    fecha: string;	
    planta_concreto: string;	
    num_empleado: string;	
    codigo_parada: string;	
    hora_inicio: string;	
    hora_termino: string;	
    observaciones_p: string;	
    codigo_parada2: string;	
    hora_inicio2: string;	
    hora_termino2: string;	
    observaciones_p2: string;	
    nivel_aceite_hidraulico: string;	
    nivel_aceite_motor: string;	
    nivel_aceite_transmision: string;	
    nivel_anticongelante: string;	
    freno_pedal: string;	
    freno_motor: string;	
    freno_estacionario: string;	
    presion_estado_neumaticos: string;	
    parabrisas: string;	
    e_bateria_conexiones: string;	
    tablero_electrico: string;	
    luces_delanteras_etc: string;	
    luces_reversa: string;	
    alarma_reversera: string;	
    torreta: string;	
    f_claxon: string;	
    estado_asientos: string;	
    cinturon_seguridad: string;	
    extintor: string;	
    botiquin: string;	
    calzas_seguridad: string;	
    herramientas_basica: string;	
    revision_muelles_bujes_perchas: string;	
    engrasar_equipo: string;	
    limpieza_general: string;	
    tanque_aceite_hidraulico: string;	
    aceite_hidraulico: string;	
    enfriador_aceite_hidraulico: string;	
    tolva_agregado_fino: string;	
    tolva_agregado_grueso: string;	
    tolva_hermetica_cemento: string;	
    bomba_agua: string;	
    deposito_agua: string;	
    flujometro_control_agua: string;	
    compuertas_dosificacion: string;	
    banda_transp_agregado: string;	
    mezclador_continuo: string;	
    seguros_mezclador: string;	
    guardas_mezclador: string;	
    luces_trabajo: string;	
    limpieza_equipo: string;	
    observaciones_reverso: string;	
    nombre_operador: string;	
    nombre_supervisor: string;	
    created_at?: string;	
    updated_a?: string;
}